Move management.py run, dump_testdata, reload_testdata to development app
