+++ evap/development/management/commands/dump_testdata.py
@@ -11,7 +11,7 @@ class Command(BaseCommand):
     requires_migrations_checks = True
 
     def handle(self, *args, **options):
-        outfile_name = os.path.join(settings.BASE_DIR, "evaluation", "fixtures", "test_data.json")
+        outfile_name = os.path.join(settings.BASE_DIR, 'development', 'fixtures', 'test_data.json')
         call_command(
             "dumpdata", "auth.group", "evaluation", "rewards", "student", "grades", "--exclude=evaluation.LogEntry", indent=2,
             output=outfile_name, natural_foreign=True, natural_primary=True)
